requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
httpx==0.27.0
streamlit==1.36.0
python-dotenv==1.0.1
pydantic==2.7.4
scikit-learn==1.5.1
pandas==2.2.2
plotly==5.22.0
